Moved .NET 4.5 methods from IRavenClient.cs to IRavenClient.Net45.cs
