+++ src/app/SharpRaven/IRavenClient.Net45.cs
@@ -2,21 +2,21 @@
 
 // Copyright (c) 2014 The Sentry Team and individual contributors.
 // All rights reserved.
-//
+// 
 // Redistribution and use in source and binary forms, with or without modification, are permitted
 // provided that the following conditions are met:
-//
+// 
 //     1. Redistributions of source code must retain the above copyright notice, this list of
 //        conditions and the following disclaimer.
-//
+// 
 //     2. Redistributions in binary form must reproduce the above copyright notice, this list of
 //        conditions and the following disclaimer in the documentation and/or other materials
 //        provided with the distribution.
-//
+// 
 //     3. Neither the name of the Sentry nor the names of its contributors may be used to
 //        endorse or promote products derived from this software without specific prior written
 //        permission.
-//
+// 
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
 // IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
 // FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
@@ -28,14 +28,64 @@
 
 #endregion
 
-using System;
-
 #if !(net40) && !(net35)
+
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+using SharpRaven.Data;
 
 namespace SharpRaven
 {
     public partial interface IRavenClient : IDisposable
     {
+        /// <summary>Captures the event.</summary>
+        /// <param name="event">The event to capture.</param>
+        /// <returns>
+        /// The <see cref="JsonPacket.EventID" /> of the successfully captured <paramref name="event" />, or <c>null</c> if it fails.
+        /// </returns>
+        Task<string> CaptureAsync(SentryEvent @event);
+
+
+        /// <summary>
+        /// Captures the <see cref="Exception" />.
+        /// </summary>
+        /// <param name="exception">The <see cref="Exception" /> to capture.</param>
+        /// <param name="message">The optional message to capture instead of the default <see cref="Exception.Message" />.</param>
+        /// <param name="level">The <see cref="ErrorLevel" /> of the captured <paramref name="exception" />. Default: <see cref="ErrorLevel.Error"/>.</param>
+        /// <param name="tags">The tags to annotate the captured <paramref name="exception" /> with.</param>
+        /// <param name="fingerprint">The custom fingerprint to identify the captured <paramref name="message" /> with.</param>
+        /// <param name="extra">The extra metadata to send with the captured <paramref name="exception" />.</param>
+        /// <returns>
+        /// The <see cref="JsonPacket.EventID" /> of the successfully captured <paramref name="exception" />, or <c>null</c> if it fails.
+        /// </returns>
+        [Obsolete("Use CaptureAsync(SentryEvent) instead.")]
+        Task<string> CaptureExceptionAsync(Exception exception,
+                                           SentryMessage message = null,
+                                           ErrorLevel level = ErrorLevel.Error,
+                                           IDictionary<string, string> tags = null,
+                                           string[] fingerprint = null,
+                                           object extra = null);
+
+
+        /// <summary>
+        /// Captures the message.
+        /// </summary>
+        /// <param name="message">The message to capture.</param>
+        /// <param name="level">The <see cref="ErrorLevel" /> of the captured <paramref name="message"/>. Default <see cref="ErrorLevel.Info"/>.</param>
+        /// <param name="tags">The tags to annotate the captured <paramref name="message"/> with.</param>
+        /// <param name="fingerprint">The custom fingerprint to identify the captured <paramref name="message" /> with.</param>
+        /// <param name="extra">The extra metadata to send with the captured <paramref name="message"/>.</param>
+        /// <returns>
+        /// The <see cref="JsonPacket.EventID"/> of the successfully captured <paramref name="message"/>, or <c>null</c> if it fails.
+        /// </returns>
+        [Obsolete("Use CaptureAsync(SentryEvent) instead.")]
+        Task<string> CaptureMessageAsync(SentryMessage message,
+                                         ErrorLevel level = ErrorLevel.Info,
+                                         IDictionary<string, string> tags = null,
+                                         string[] fingerprint = null,
+                                         object extra = null);
     }
 }
